Re-add changes lost due to a bad merge Change workspace rules to packages/shared/*, pacakges/*-app and packages/veto- Add @veto/utils package. It currently hold `config.ts` Add unit test for 'getDockerVersion.ts` Remove direct jest dependency and (we used ts-jest's version instead) Rename build.sh to setup.sh to follow current convention (setup, start, stop) Use setup script for the `yarn cyphernode:setup` command Add packages/ structure to README.md
