+++ packages/shared/utils/.gitignore
@@ -1,0 +1,2 @@
+# Dependency directories
+node_modules/
