--- packages/veto-backend/README.md
@@ -1,1 +1,0 @@
-To be reviewed since Next app is not used anymore
