+++ packages/veto-backend/package.json
@@ -4,7 +4,7 @@
   "workspaces": [
     "packages/*"
   ],
-  "description": "Veriphy software",
+  "description": "Veto software",
   "main": "src/index.ts",
   "repository": "git@github.com:Veriphi/veto.git",
   "author": "Veriphi",
@@ -16,6 +16,7 @@
     "start": "ts-node src/index.ts"
   },
   "dependencies": {
+    "@veto/utils": "0.0.0",
     "axios": "^0.19.1",
     "express": "^4.17.1",
     "express-http-proxy": "^1.6.0"
